==> 4_evaluate_diagnostic_datasets.ipynb <==
# Evaluate Diagnostic Datasets

- Use the previously created diagnostic datasets.
- Evaluate every diagnostic dataset with both ranking models (BM25 and neural).
- Load all results with pandas and display each dataset's correlation values and whether the values are significant.

from src import get_data, diagnostic, scorer, util
import pandas as pd
pd.set_option('display.max_colwidth', None)
pd.set_option('display.max_rows', None)

bm25_scorer = scorer.BM25Scorer()

for dataobj in get_data.CqaDupStackCollector().iter_datasets():
    for qset in dataobj.list_qsets():
        for diag_dataset_name in dataobj.list_probes(qset):
            diag_dataset = diagnostic.DiagnosticDataset(dataobj, qset, diag_dataset_name)
            if not diag_dataset.has_scored('bm25'):
                bm25_scorer.score_dataset(diag_dataset, dataobj, True)

neural_cqa_scorer = scorer.NeuralScorer('GPL/cqadupstack-msmarco-distilbert-gpl')

for dataobj in get_data.CqaDupStackCollector().iter_datasets():
    for qset in dataobj.list_qsets():
        for diag_dataset_name in dataobj.list_probes(qset):
            diag_dataset = diagnostic.DiagnosticDataset(dataobj, qset, diag_dataset_name)
            if not diag_dataset.has_scored('neural'):
                neural_cqa_scorer.score_dataset(diag_dataset, dataobj, True)

for dataobj in get_data.CqaDupStackCollector().iter_datasets():
    name = dataobj.path.name
    correlation_df, significant_df = util.create_results_dfs(dataobj, name)
    print(name)
    display(correlation_df)
    display(significant_df)
